third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/breakpoint-manager-expected.txt

Tests BreakpointManager class.


Running: testNavigation
  Created breakpoints manager
  Dumping Storage
    a.js:10 enabled:true condition:foo == bar
    a.js:20 enabled:false condition:
    b.js:3 enabled:true condition:
  Adding script: a.js
  Adding UISourceCode: a.js
    breakpointAdded(a.js, 10, 0, foo == bar, true)
    debuggerModel.setBreakpoint(a.js:10:foo == bar)
    breakpointAdded(a.js, 20, 0, , false)
    breakpointRemoved(a.js, 10, 0)
    breakpointAdded(a.js, 10, 0, foo == bar, true)
    Location created: a.js:10
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (10, 0)
      Location: (20, 0)

  Navigating to B.
  Resetting debugger.
  Resetting workspace.
    breakpointRemoved(a.js, 10, 0)
    Location disposed: a.js:10
    breakpointRemoved(a.js, 20, 0)
  Adding script: b.js
  Adding UISourceCode: b.js
    breakpointAdded(b.js, 3, 0, , true)
    debuggerModel.setBreakpoint(b.js:3:)
    breakpointRemoved(b.js, 3, 0)
    breakpointAdded(b.js, 3, 0, , true)
    Location created: b.js:3
  Dumping Breakpoint Locations
    UISourceCode (url='b.js', uri='b.js')
      Location: (3, 0)

  Navigating back to A.
  Resetting debugger.
  Resetting workspace.
    breakpointRemoved(b.js, 3, 0)
    Location disposed: b.js:3
  Resolving provisional breakpoint.
  Adding script: a.js
    breakpointAdded(debugger:///VMa.js a.js, 10, 0, foo == bar, true)
    Location created: a.js:10
  Adding UISourceCode: a.js
    breakpointRemoved(debugger:///VMa.js a.js, 10, 0)
    breakpointAdded(a.js, 10, 0, foo == bar, true)
    breakpointAdded(a.js, 20, 0, , false)
    breakpointRemoved(a.js, 10, 0)
    breakpointAdded(a.js, 10, 0, foo == bar, true)
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (10, 0)
      Location: (20, 0)
  Dumping Storage
    a.js:10 enabled:true condition:foo == bar
    a.js:20 enabled:false condition:
    b.js:3 enabled:true condition:
  Resetting breakpoint manager
    breakpointRemoved(a.js, 20, 0)
    breakpointRemoved(a.js, 10, 0)
    Location disposed: a.js:10
    debuggerModel.removeBreakpoint(a.js:10)
    debuggerModel.removeBreakpoint(b.js:3)
  Dumping Breakpoint Locations

Running: testProvisionalBreakpointsResolve
  Created breakpoints manager
  Dumping Storage
    a.js:10 enabled:true condition:foo == bar
  Adding script: a.js
  Adding UISourceCode: a.js
    breakpointAdded(a.js, 10, 0, foo == bar, true)
    debuggerModel.setBreakpoint(a.js:10:foo == bar)
    breakpointRemoved(a.js, 10, 0)
    breakpointAdded(a.js, 10, 0, foo == bar, true)
    Location created: a.js:10
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (10, 0)
  Resetting debugger.
  Resetting workspace.
    breakpointRemoved(a.js, 10, 0)
    Location disposed: a.js:10
  Resolving provisional breakpoint.
  Adding script: a.js
    breakpointAdded(debugger:///VMa.js a.js, 11, 0, foo == bar, true)
    Location created: a.js:11
  Dumping Breakpoint Locations
    UISourceCode (url='debugger:///VMa.js a.js', uri='debugger:///VMa.js a.js')
      Location: (11, 0)
  Dumping Storage
    a.js:10 enabled:true condition:foo == bar
  Resetting breakpoint manager
    breakpointRemoved(debugger:///VMa.js a.js, 11, 0)
    Location disposed: a.js:11
    debuggerModel.removeBreakpoint(a.js:10)
  Dumping Breakpoint Locations

Running: testBreakpointInCollectedReload
  Created breakpoints manager
  Dumping Storage

  Adding file without script:
  Adding UISourceCode: a.js

  Setting breakpoint:
  Setting breakpoint at a.js:10:0 enabled:true condition:
    breakpointAdded(a.js, 10, 0, , true)
    debuggerModel.setBreakpoint(a.js:10:)
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (10, 0)

  Reloading:
  Resetting debugger.
  Resetting workspace.
    breakpointRemoved(a.js, 10, 0)

  Adding file with script:
  Adding script: a.js
  Adding UISourceCode: a.js
    breakpointAdded(a.js, 10, 0, , true)

  Emulating breakpoint resolved event:
    breakpointRemoved(a.js, 10, 0)
    breakpointAdded(a.js, 10, 0, , true)
    Location created: a.js:10

  Make sure we don't do any unnecessary breakpoint actions:
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (10, 0)
  Dumping Storage
    a.js:10 enabled:true condition:
  Resetting breakpoint manager
    breakpointRemoved(a.js, 10, 0)
    Location disposed: a.js:10
    debuggerModel.removeBreakpoint(a.js:10)
  Dumping Breakpoint Locations
